orb-workflow/pyproject.toml
[tool.poetry]
name = "orb-workflow"
version = "0.1.0"
description = "LangGraph workflow for Orb billing automation"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.1.10"
langchain-fireworks = "^0.1"
python-dotenv = "^1.0.1"
fireworks-ai = "^0.13.0"
watchdog = "^4.0.0"
httpx = "^0.27.0"
pydantic = "^2.0"
pyjwt = "^2.8.0"
openai = "^1.35.10"

# Add the local toolkit as a development dependency
arcade-orb-toolkit = {path = "../arcade/orb_toolkit", develop = true}
arcadepy = "^1.0.0"
google-api-python-client = "^2.167.0"
google-auth = "^2.39.0"
pypdf2 = "^3.0.1"
slack_sdk = "^3.29.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 